Error handling done:
1. There are cases where there would be legit errors like when parsing 1.0.0.
In this case now the API will throw error along with false.

2. There are times when the info passed in the items is incomplete or is of the wrong type.
In this case API now has the ability to give you debug reason for why you got false.

3. This removes all the panic from this lib.

Adding proper error on operations. Adding error in the parser and persisting it. Operation errors are able to give us the ability to get debug info